+++ src/main/java/com/uwetrottmann/tmdb2/services/ConfigurationService.java
@@ -2,6 +2,7 @@ package com.uwetrottmann.tmdb2.services;
 
 import com.uwetrottmann.tmdb2.entities.Configuration;
 import com.uwetrottmann.tmdb2.entities.Jobs;
+import com.uwetrottmann.tmdb2.entities.TmdbLanguage;
 import java.util.List;
 import retrofit2.Call;
 import retrofit2.http.GET;
@@ -28,5 +29,11 @@ public interface ConfigurationService {
      */
     @GET("configuration/jobs")
     Call<List<Jobs>> jobs();
+
+    @GET("configuration/languages")
+    Call<List<TmdbLanguage>> languages();
+
+    @GET("configuration/primary_translations")
+    Call<List<String>> primary_translations();
 
 }